Refactor Input components and CreateDialogPatch screen for improved readability and maintainability through layout redesign and styling.
