+++ Components/Input/common/InputLocalization.razor
@@ -2,39 +2,26 @@
 @using System.Diagnostics
 @using System.Diagnostics.CodeAnalysis
 @using Altinn.ApiClients.Dialogporten.Features.V1
-<div>
-    <table>
-        <thead>
-        @if ((Value ?? throw new InvalidOperationException()).Any())
-        {
-            <tr>
-                <th>Language Code</th>
-                <th>Text</th>
-                <th>Actions</th>
-            </tr>
-        }
-        </thead>
-        <tbody>
-        @foreach (var localization in Value!)
-        {
-            <tr>
-                <td>
-                    <InputText @bind-Value="localization.LanguageCode" placeholder="Language Code"/>
-                </td>
-                <td>
-                    <InputText @bind-Value="localization.Value" placeholder="Localized Value"/>
-                </td>
-                <td>
-                    <button type="button" @onclick="() => Remove(localization)">Remove</button>
-                </td>
-            </tr>
-        }
-        </tbody>
-    </table>
 
-    <div>
-        <button type="button" @onclick="Add">Add Localization</button>
+<div class="localization-component">
+    <div class="localization-list">
+        @foreach (var localization in (Value ?? throw new InvalidOperationException()))
+        {
+            <div class="localization-item">
+                <div class="localization-fields">
+                    <div class="form-group">
+                        <InputText class="form-input" @bind-Value="localization.LanguageCode" placeholder="Language Code (e.g., en, no)"/>
+                    </div>
+                    <div class="form-group">
+                        <InputText class="form-input" @bind-Value="localization.Value" placeholder="Localized Text"/>
+                    </div>
+                </div>
+                <button type="button" class="btn btn-remove btn-small" @onclick="() => Remove(localization)">Remove</button>
+            </div>
+        }
     </div>
+
+    <button type="button" class="btn btn-add btn-small" @onclick="Add">Add Localization</button>
 </div>
 
 @code{
@@ -67,3 +54,38 @@
         ValueChanged.InvokeAsync(Value);
     }
 }
+
+<style>
+    .localization-component {
+        margin-bottom: 0.5rem;
+    }
+
+    .localization-list {
+        display: flex;
+        flex-direction: column;
+        gap: 0.5rem;
+        margin-bottom: 0.5rem;
+    }
+
+    .localization-item {
+        display: flex;
+        gap: 0.5rem;
+        align-items: center;
+        background: #f8f9fa;
+        padding: 0.5rem;
+        border-radius: 4px;
+    }
+
+    .localization-fields {
+        display: flex;
+        gap: 0.5rem;
+        flex: 1;
+        flex-wrap: wrap;
+    }
+
+    .localization-fields .form-group {
+        flex: 1;
+        min-width: 150px;
+        margin-bottom: 0;
+    }
+</style>
